+++ src/views/components/GameCard/GameCard.tsx
@@ -1,7 +1,6 @@
 import style from './GameCard.module.scss'
 import { IGameCard } from './IGameCard'
-import { CardHeader, CardScore } from '@views/components'
-import { Game } from '@data/types'
+import { CardHeader, CardMeta, CardScore } from '@views/components'
 
 
 export { GameCard }
@@ -13,26 +12,5 @@ function GameCard({ game, lazyLoad }: IGameCard) {
 			<CardScore game={game}/>
 			<CardMeta game={game}/>
 		</article>
-	)
-}
-
-interface ICardMeta {
-	game: Game
-}
-
-function CardMeta({ game }: ICardMeta) {
-	const gameTags = game.tags.map(t => t.name).slice(0, 3).join(', ')
-
-	return (
-		<>
-			<p className={style.Tags}>
-				tags: {gameTags}
-			</p>
-
-			<p className={style.Genres}>
-				<span className="sr-only">Genre:</span>
-				{game.genres[0]?.name ?? 'Not listed'}
-			</p>
-		</>
 	)
 }
